String reverse

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -11,7 +11,7 @@ void rev_string(char *s)
 	int i;
 	char tmp;
 
-	for (i = 0; i < strlen(s) / 2; i++)
+	for (i = 0; i < (int)  strlen(s) / 2; i++)
 	{
 		tmp = s[i];
 		s[i] = s[strlen(s) - 1 - i];
